+++ api/rodos-debug.h
@@ -55,7 +55,7 @@ extern const char* rodosErrorMsg;
  * @param[in/out]  outBuf buffer to write string 
  * @return  outBuf buffer to write string
  */
-char* formatBinary(long val, int len, char* outBuf);
+char* formatBinary(uint32_t val, size_t len, char* outBuf);
 
 /** Printf protected with a semaphore: Thread safe 
      Warning: %3.9f means 3 digits then 9 digits
@@ -65,7 +65,7 @@ void PRINTF(const char* fmt, ...) __attribute__((__format__(__printf__,1,2)));
 
 void SPRINTF(char* dest, const char* fmt, ...) __attribute__((__format__(__printf__,2,3)));
 
-int SNPRINTF(char* dest, size_t size, const char* fmt, ...) __attribute__((__format__(__printf__,3,4)));
+size_t SNPRINTF(char* dest, size_t size, const char* fmt, ...) __attribute__((__format__(__printf__,3,4)));
 
 /** id shall have only one bit set, prints only if (id & printfMask)  **/
 void PRINTF_CONDITIONAL(uint32_t id, const char* fmt, ...) __attribute__((__format__(__printf__,2,3)));
