+++ src/bare-metal/sf2/rodos/hw_tcp.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "lwip/tcp.h"
+#include <stddef.h>
 
 err_t tcp_recv_func(void* arg, struct tcp_pcb* tpcb, struct pbuf* p, err_t);
 err_t tcp_accept_func(void* arg, struct tcp_pcb* newpcb, err_t err);
@@ -12,8 +13,8 @@ class TCPBase {
   public:
     virtual ~TCPBase() { tcp_abort(pcb); }
 
-    virtual int sendData(void* buf, int len);
-    virtual int getData(void* buf, int maxLen);
+    virtual int sendData(void* buf, size_t len);
+    virtual int getData(void* buf, size_t maxLen);
 
   protected:
     pbuf*    recvBuf   = nullptr;
@@ -21,7 +22,7 @@ class TCPBase {
     int      errorCode = -1;
 
   protected:
-    int readPos = 0;
+    size_t readPos = 0;
 
     friend err_t ::tcp_accept_func(void* arg, struct tcp_pcb* newpcb, err_t err);
     friend err_t ::tcp_recv_func(void* arg, struct tcp_pcb* tpcb, struct pbuf* p, err_t);
@@ -34,8 +35,8 @@ class HW_TCPServer : public TCPBase {
 
     bool listen(const long portNr);
     bool acceptNewConnection();
-    int  sendData(void* buf, int len) override;
-    int  getData(void* buf, int maxLen) override;
+    int  sendData(void* buf, size_t len) override;
+    int  getData(void* buf, size_t maxLen) override;
     int  getErrorCode() const;
 
     tcp_pcb* listenPcb = nullptr;
@@ -46,8 +47,8 @@ class HW_TCPClient : public TCPBase {
     ~HW_TCPClient();
 
     bool reopen(const long portNr, const char* hostname = "localhost");
-    int  sendData(void* buf, int len) override;
-    int  getData(void* buf, int maxLen) override;
+    int  sendData(void* buf, size_t len) override;
+    int  getData(void* buf, size_t maxLen) override;
     int  getErrorCode() const;
 };
